Update analytics integrations to handle MV flags
