+++ api/integrations/mixpanel/mixpanel.py
@@ -1,9 +1,14 @@
 import json
 import logging
+import typing
 
 import requests
 
 from integrations.common.wrapper import AbstractBaseIdentityIntegrationWrapper
+
+if typing.TYPE_CHECKING:
+    from environments.identities.models import Identity
+    from features.models import FeatureState
 
 logger = logging.getLogger(__name__)
 
@@ -28,17 +33,19 @@ class MixpanelWrapper(AbstractBaseIdentityIntegrationWrapper):
         )
         logger.debug("Sent event to Mixpanel. Body code was: %s" % response.content)
 
-    def generate_user_data(self, user_id, feature_states):
+    def generate_user_data(
+        self, identity: "Identity", feature_states: typing.List["FeatureState"]
+    ) -> dict:
         feature_properties = {}
 
         for feature_state in feature_states:
-            value = feature_state.get_feature_state_value()
+            value = feature_state.get_feature_state_value(identity=identity)
             feature_properties[feature_state.feature.name] = (
                 value if (feature_state.enabled and value) else feature_state.enabled
             )
 
         return {
             "$token": self.api_key,
-            "$distinct_id": user_id,
+            "$distinct_id": identity.identifier,
             "$set": feature_properties,
         }
