+++ api/integrations/segment/segment.py
@@ -1,7 +1,12 @@
 import logging
+import typing
 
 from analytics.client import Client as SegmentClient
 from integrations.common.wrapper import AbstractBaseIdentityIntegrationWrapper
+
+if typing.TYPE_CHECKING:
+    from environments.identities.models import Identity
+    from features.models import FeatureState
 
 logger = logging.getLogger(__name__)
 
@@ -14,7 +19,9 @@ class SegmentWrapper(AbstractBaseIdentityIntegrationWrapper):
         self.analytics.identify(**data)
         logger.debug(f"Sent event to Segment.")
 
-    def generate_user_data(self, user_id, feature_states):
+    def generate_user_data(
+        self, identity: "Identity", feature_states: typing.List["FeatureState"]
+    ) -> dict:
         feature_properties = {}
 
         for feature_state in feature_states:
@@ -24,6 +31,6 @@ class SegmentWrapper(AbstractBaseIdentityIntegrationWrapper):
             )
 
         return {
-            "user_id": user_id,
+            "user_id": identity.identifier,
             "traits": feature_properties,
         }
