+++ api/tests/unit/integrations/amplitude/test_unit_amplitude.py
@@ -1,29 +1,34 @@
 import pytest
 
+from environments.identities.models import Identity
 from environments.models import Environment
 from features.models import Feature, FeatureState
-from integrations.segment.segment import SegmentWrapper
+from integrations.amplitude.amplitude import (
+    AMPLITUDE_API_URL,
+    AmplitudeWrapper,
+)
 from organisations.models import Organisation
 from projects.models import Project
 
 
-def test_segment_initialized_correctly():
+def test_amplitude_initialized_correctly():
     # Given
     api_key = "123key"
 
     # When initialized
-    segment_wrapper = SegmentWrapper(api_key=api_key)
+    amplitude_wrapper = AmplitudeWrapper(api_key=api_key)
 
     # Then
-    assert segment_wrapper.analytics.write_key == api_key
+    expected_url = f"{AMPLITUDE_API_URL}/identify"
+    assert amplitude_wrapper.url == expected_url
 
 
 @pytest.mark.django_db
-def test_segment_when_generate_user_data_with_correct_values_then_success():
+def test_amplitude_when_generate_user_data_with_correct_values_then_success():
     # Given
     api_key = "123key"
-    user_id = "user123"
-    segment_wrapper = SegmentWrapper(api_key=api_key)
+    identity = Identity(identifier="user123")
+    amplitude_wrapper = AmplitudeWrapper(api_key=api_key)
 
     organisation = Organisation.objects.create(name="Test Org")
     project = Project.objects.create(name="Test Project", organisation=organisation)
@@ -34,12 +39,13 @@ def test_segment_when_generate_user_data_with_correct_values_then_success():
     feature_states = FeatureState.objects.filter(feature=feature)
 
     # When
-    user_data = segment_wrapper.generate_user_data(
-        user_id=user_id, feature_states=feature_states
+    user_data = amplitude_wrapper.generate_user_data(
+        identity=identity, feature_states=feature_states
     )
 
     # Then
     feature_properties = {}
+
     for feature_state in feature_states:
         value = feature_state.get_feature_state_value()
         feature_properties[feature_state.feature.name] = (
@@ -47,8 +53,8 @@ def test_segment_when_generate_user_data_with_correct_values_then_success():
         )
 
     expected_user_data = {
-        "user_id": user_id,
-        "traits": feature_properties,
+        "user_id": identity.identifier,
+        "user_properties": feature_properties,
     }
 
     assert expected_user_data == user_data
